refactor: Rename `lib/plugins/aws/lib/normalizeFiles.js`
