+++ lib/plugins/aws/deploy/lib/checkForChanges.js
@@ -6,7 +6,7 @@ const crypto = require('crypto');
 const globby = require('globby');
 const BbPromise = require('bluebird');
 const _ = require('lodash');
-const normalizeFiles = require('../../lib/normalizeFiles');
+const normalizeFiles = require('../../lib/normalize-files');
 const ServerlessError = require('../../../../serverless-error');
 const { log } = require('@serverless/utils/log');
