+++ lib/plugins/aws/deploy/lib/uploadArtifacts.js
@@ -7,7 +7,7 @@ const path = require('path');
 const crypto = require('crypto');
 const limit = require('ext/promise/limit').bind(Promise);
 const filesize = require('filesize');
-const normalizeFiles = require('../../lib/normalizeFiles');
+const normalizeFiles = require('../../lib/normalize-files');
 const getLambdaLayerArtifactPath = require('../../utils/getLambdaLayerArtifactPath');
 const ServerlessError = require('../../../../serverless-error');
 const { progress, log } = require('@serverless/utils/log');
