+++ lib/plugins/aws/lib/normalize-files.js
