+++ test/unit/lib/plugins/aws/deploy/lib/checkForChanges.test.js
@@ -9,7 +9,7 @@ const globby = require('globby');
 const sandbox = require('sinon');
 const chai = require('chai');
 const proxyquire = require('proxyquire');
-const normalizeFiles = require('../../../../../../../lib/plugins/aws/lib/normalizeFiles');
+const normalizeFiles = require('../../../../../../../lib/plugins/aws/lib/normalize-files');
 const AwsProvider = require('../../../../../../../lib/plugins/aws/provider');
 const AwsDeploy = require('../../../../../../../lib/plugins/aws/deploy/index');
 const Serverless = require('../../../../../../../lib/serverless');
