+++ test/unit/lib/plugins/aws/deploy/lib/uploadArtifacts.test.js
@@ -8,7 +8,7 @@ const fse = require('fs-extra');
 const path = require('path');
 const chai = require('chai');
 const proxyquire = require('proxyquire');
-const normalizeFiles = require('../../../../../../../lib/plugins/aws/lib/normalizeFiles');
+const normalizeFiles = require('../../../../../../../lib/plugins/aws/lib/normalize-files');
 const AwsProvider = require('../../../../../../../lib/plugins/aws/provider');
 const AwsDeploy = require('../../../../../../../lib/plugins/aws/deploy/index');
 const Serverless = require('../../../../../../../lib/serverless');
